MAINT: Restructure and split coordinate transformation and affine correction.
